+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Changed
+
+-   Email verification endpoints will now clear the session if called by a deleted/unknown user
+
 ## [12.0.0] - 2022-09-14
 
 ### Bug fix: